# docs/complete-api-swagger.yaml
openapi: 3.0.0
info:
  title: 🛒 Ghanshyam Murti Bhandar - Complete Ecommerce API
  description: |
    # 🚀 Enterprise-Grade Ecommerce Platform API
    
    **239+ API Endpoints** covering all aspects of modern ecommerce for Android integration.
    
    ## 🔥 Complete Feature Set
    - **🔐 Authentication** - JWT-based secure authentication (4 endpoints)
    - **📦 Products** - Complete product management (12 endpoints)
    - **🏷️ Categories** - Category management (8 endpoints)
    - **🛒 Shopping Cart** - Real-time cart management (6 endpoints)
    - **❤️ Wishlist** - Save favorite products (5 endpoints)
    - **📍 Addresses** - Multi-address support (6 endpoints)
    - **📋 Orders** - Complete order lifecycle (15 endpoints)
    - **💳 Payments** - Razorpay, COD, Wallet (10 endpoints)
    - **🚚 Shipping** - Shiprocket integration (12 endpoints)
    - **💰 Wallet** - Digital wallet system (8 endpoints)
    - **🎫 Coupons** - Discount system (10 endpoints)
    - **⭐ Reviews** - Rating system (8 endpoints)
    - **👥 Users** - User management (10 endpoints)
    - **🏢 Admin** - Admin dashboard (25 endpoints)
    - **📊 Analytics** - Business intelligence (15 endpoints)
    - **📁 File Upload** - Image management (5 endpoints)
    - **🔔 Notifications** - Real-time updates (6 endpoints)
    - **🧾 Invoices** - GST compliance (8 endpoints)
    - **📦 Inventory** - Stock management (12 endpoints)
    - **🏪 Suppliers** - Vendor management (10 endpoints)
    
    ## 🎯 Android Integration Ready
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Get token from `/auth/login`.
        
        **Usage:** `Authorization: Bearer <your-jwt-token>`
        
        **Example:** `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Request success status
          example: true
        message:
          type: string
          description: Response message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-07-25T14:30:00Z"
      required:
        - success
        - message

    # Error Response
...
paths:
...
  /orders/{id}:
    get:
      tags:
        - 📋 Orders
      summary: 🔍 Get Order Details
      description: |
        Get detailed information about a specific order.

        **Android Integration:**
        - Use for order detail screen
        - Show order tracking
        - Display payment status
        - Enable order actions (cancel, return)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID
          example: "60f1b2b3c4d5e6f7g8h9i0j6"
      responses:
        '200':
          description: ✅ Order details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          order:
                            type: object
                            properties:
                              _id:
                                type: string
                              orderNumber:
                                type: string
                              status:
                                type: string
                              items:
                                type: array
                                items:
                                  type: object
                              pricing:
                                type: object
                              paymentInfo:
                                type: object
                              shippingAddress:
                                type: object
                              tracking:
                                type: object
                                properties:
                                  awbCode:
                                    type: string
                                  courierName:
                                    type: string
                                  trackingUrl:
                                    type: string
                              timeline:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    status:
                                      type: string
                                    timestamp:
                                      type: string
                                      format: date-time
                                    description:
                                      type: string
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Order not found
        '500':
          description: ❌ Internal server error

  # 💳 PAYMENT ENDPOINTS
...